requirements.txt
aiohttp==3.10.3
aiosignal==1.3.1
attrs==24.2.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==2.0.12
click==8.1.7
dash==2.18.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==3.0.3
Flask-Caching==1.10.1
gunicorn==20.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
multidict==6.0.5
numpy==2.0.2
pandas==2.2.3
plotly==5.24.1
python-dateutil==2.9.0.post0
python-dotenv==0.19.0
requests==2.26.0
retrying==1.3.4
six==1.16.0
SQLAlchemy==2.0.23
tenacity==8.3.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.20
Werkzeug==3.0.4
yarl==1.9.4